refactor: Rename ResendPassword > ResendPasswordPage
